refactor(catalog): TrackDetailController -> TrackDetailViewModel (file, class, provider)
