+++ apps/mobile_app/lib/features/catalog/presentation/viewmodels/learn_view_model.dart
