+++ apps/mobile_app/lib/features/catalog/presentation/viewmodels/track_detail_view_model.dart
